Add new cache strategy: `Dummy`; Some updates and fixes
