+++ .travis.yml
@@ -10,7 +10,6 @@ dist: trusty
 
 services:
   - postgresql
-  - redis-server
 
 addons:
   postgresql: "9.5"
