+++ backend/cache/redis.go
@@ -1,9 +1,6 @@
 package cache
 
 import (
-	"strconv"
-	"strings"
-
 	log "github.com/Sirupsen/logrus"
 	"gopkg.in/redis.v5"
 
@@ -18,18 +15,8 @@ type RedisCache struct {
 	config config.CacheConfigProvider
 }
 
-type Fields []interface{}
-
 // Creates new CacheProvider from Redis client
-func NewRedisCache(client *redis.Client, config config.CacheConfigProvider) CacheProvider {
-	return &RedisCache{
-		client: client,
-		config: config,
-	}
-}
-
-// Establishes new connection to Redis
-func NewRedisConnection(port string) *redis.Client {
+func NewRedisCache(port string, config config.CacheConfigProvider) CacheProvider {
 	// TODO: read user data, host and port from config file
 	client := redis.NewClient(&redis.Options{
 		Addr:     "localhost:" + port,
@@ -42,7 +29,10 @@ func NewRedisConnection(port string) *redis.Client {
 		log.WithError(err).Fatal("Error connecting to cache instance.")
 	}
 
-	return client
+	return &RedisCache{
+		client: client,
+		config: config,
+	}
 }
 
 // Set `value` for specified `key`
@@ -58,7 +48,7 @@ func (cache *RedisCache) Set(key string, value interface{}) {
 
 // Set `value` for specified key where key is created by hashing `fields`
 func (cache *RedisCache) SetWithFields(fields Fields, value interface{}) {
-	cache.Set(cache.convertFieldsToKey(fields), value)
+	cache.Set(convertFieldsToKey(fields), value)
 }
 
 // Get value for specified `key`
@@ -74,7 +64,7 @@ func (cache *RedisCache) Get(key string) (interface{}, bool) {
 
 // Get value for specified key where key is created by hashing `fields`
 func (cache *RedisCache) GetWithFields(fields Fields) (interface{}, bool) {
-	return cache.Get(cache.convertFieldsToKey(fields))
+	return cache.Get(convertFieldsToKey(fields))
 }
 
 // Delete value for specific `key`
@@ -87,23 +77,5 @@ func (cache *RedisCache) Delete(key string) {
 
 // Delete value for specific key where key is created by hashing `fields`
 func (cache *RedisCache) DeleteWithFields(fields Fields) {
-	cache.Delete(cache.convertFieldsToKey(fields))
-}
-
-// Joins multiple fields to single key
-func (cache *RedisCache) convertFieldsToKey(fields Fields) string {
-	var stringFields []string
-	for _, field := range fields {
-		switch field.(type) {
-		case string:
-			stringFields = append(stringFields, field.(string))
-		case int:
-			stringFields = append(stringFields, strconv.FormatInt(int64(field.(int)), 10))
-		case int32:
-			stringFields = append(stringFields, strconv.FormatInt(int64(field.(int32)), 10))
-		case int64:
-			stringFields = append(stringFields, strconv.FormatInt(int64(field.(int64)), 10))
-		}
-	}
-	return strings.Join(stringFields, "_")
+	cache.Delete(convertFieldsToKey(fields))
 }
